This patch changes the mozilla isp db data parser and data structure to make use of the types already defined for use in the settings exporter. Further extends those types with AuthenticationType and ServerType, also extends ConnectionSecurity. Changing the Stores to make use of the types. Each store has a constructor only taking the settings for argument, not an entire account.
Note: The stores were only changed to compile a function, with the new
enums some code could be simplified.

Note: This patch BREAKS SETTINGS EXPORT. The problem lies in the extra's
in ServerSettings. Using a Bundle instead of a Map<String,String>
greatly simplifies a lot of the code but a new solution must be found to
import and export it.